Revert "Fixup RedundantOptionInspection and OptionBaseOneInspection"
This reverts commit ad3b5efe5d4072fb63056ecb6ecaeed38113c235.
