+++ Rubberduck.Inspections/Rubberduck.Inspections.csproj
@@ -58,7 +58,6 @@
     <Compile Include="Concrete\ApplicationWorksheetFunctionInspection.cs" />
     <Compile Include="Concrete\AssignedByValParameterInspection.cs" />
     <Compile Include="Concrete\LineLabelNotUsedInspection.cs" />
-    <Compile Include="Concrete\OptionBaseOneInspection.cs" />
     <Compile Include="ParseTreeListeners\AttributeAnnotationListener.cs" />
     <Compile Include="Concrete\ConstantNotUsedInspection.cs" />
     <Compile Include="Concrete\DefaultProjectNameInspection.cs" />
@@ -93,6 +92,7 @@
     <Compile Include="Concrete\ObsoleteGlobalInspection.cs" />
     <Compile Include="Concrete\ObsoleteLetStatementInspection.cs" />
     <Compile Include="Concrete\ObsoleteTypeHintInspection.cs" />
+    <Compile Include="Concrete\OptionBaseInspection.cs" />
     <Compile Include="Concrete\RedundantOptionInspection.cs" />
     <Compile Include="Concrete\OptionExplicitInspection.cs" />
     <Compile Include="Concrete\ParameterCanBeByValInspection.cs" />
@@ -104,7 +104,6 @@
     <Compile Include="QuickFixes\ApplicationWorksheetFunctionQuickFix.cs" />
     <Compile Include="QuickFixes\AssignedByValParameterMakeLocalCopyQuickFix.cs" />
     <Compile Include="QuickFixes\ChangeDimToPrivateQuickFix.cs" />
-    <Compile Include="QuickFixes\RemoveRedundantOptionQuickFix.cs" />
     <Compile Include="QuickFixes\SpecifyExplicitByRefModifierQuickFix.cs" />
     <Compile Include="QuickFixes\ChangeProcedureToFunctionQuickFix.cs" />
     <Compile Include="QuickFixes\ConvertToProcedureQuickFix.cs" />
@@ -121,7 +120,7 @@
     <Compile Include="QuickFixes\RemoveEmptyIfBlockQuickFix.cs" />
     <Compile Include="QuickFixes\RemoveExplicitCallStatmentQuickFix.cs" />
     <Compile Include="QuickFixes\RemoveExplicitLetStatementQuickFix.cs" />
-    <Compile Include="QuickFixes\RemoveOptionBaseOneStatementQuickFix.cs" />
+    <Compile Include="QuickFixes\RemoveOptionBaseStatementQuickFix.cs" />
     <Compile Include="QuickFixes\RemoveExplicitByRefModifierQuickFix.cs" />
     <Compile Include="QuickFixes\RemoveTypeHintsQuickFix.cs" />
     <Compile Include="QuickFixes\RemoveUnassignedIdentifierQuickFix.cs" />
